script/requirements.txt
Flask==2.2.2
sqlalchemy==2.0.29
Werkzeug==2.2.2
requests==2.26.0
pytest==6.2.5
argon2-cffi==23.1.0